int32_t id -> uint64_t - partial

+++ eVSSIM/FTL_SOURCE/PAGE_MAP/ftl_adr_strategy.c
@@ -1,14 +1,14 @@
 #include "ftl_adr_strategy.h"
 
-int _FTL_READ(int32_t sector_nb, unsigned int offset, unsigned int length)
+int _FTL_READ(uint64_t sector_nb, unsigned int offset, unsigned int length)
 {
     return _FTL_READ_ADR(sector_nb, length);
 }
 
-int _FTL_READ_ADR(int32_t sector_nb, unsigned int length)
+int _FTL_READ_ADR(uint64_t sector_nb, unsigned int length)
 {
 #ifdef FTL_DEBUG
-	printf("[%s] Start: sector_nb %d length %u\n",__FUNCTION__, sector_nb, length);
+	printf("[%s] Start: sector_nb %ld length %u\n",__FUNCTION__, sector_nb, length);
 #endif
 
 	if (sector_nb + length > SECTOR_NB)
@@ -19,10 +19,10 @@ int _FTL_READ_ADR(int32_t sector_nb, unsigned int length)
 
 	int32_t lpn;
 	int32_t ppn;
-	int32_t lba = sector_nb;
+	uint64_t lba = sector_nb;
 	unsigned int remain = length;
-	unsigned int left_skip = sector_nb % SECTORS_PER_PAGE;
-	unsigned int right_skip;
+	unsigned long left_skip = sector_nb % SECTORS_PER_PAGE;
+	unsigned long right_skip;
 	unsigned int read_sects;
 
 	unsigned int ret = FAIL;
